pyproject.toml
[tool.poetry]
name = "mynetwork"
version = "0.1.2"
description = ""
authors = ["Yogesh Sharma <yks0000@gmail.com>"]
readme = "README.md"
packages = [{include = "mynetwork", from = "src"}]

[tool.poetry.scripts]
mynetwork = "mynetwork.cli:main"


[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
requests = "^2.28.1"
speedtest-cli = "^2.1.3"
hurry-filesize = "^0.9"
rich = "^12.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
